+++ .gitignore
@@ -1,6 +1,5 @@
 test.py
 .idea/
-*.json
 *.txt
 /.vscode
 /.vs
