+++ cursor-recorder.json
@@ -1,0 +1,1 @@
+{"refreshAmount": 60, "times": [0.017007, 0.3761771, 0.3931832, 0.445, 0.4612131, 0.479, 0.4952252, 0.5132326, 0.530237, 0.5472435, 0.5642498, 0.685, 0.7012981, 0.7183043, 0.7353099, 0.753, 0.7693222, 0.7872665, 0.8043349, 0.8213407, 0.8383457, 0.855353, 0.872359, 0.8893639, 0.9073119, 0.9243766, 0.942384, 0.960343, 0.9773584, 0.9953417, 1.0124089, 1.029415, 1.0464206, 1.0644291, 1.081433, 1.0984427, 1.1164455, 1.1344519, 1.1514576, 1.17, 1.1864212, 1.2034764, 1.2204824, 1.2374884, 1.2554949, 1.272501, 1.2895071, 1.3074659, 1.3244772, 1.3414804, 1.3585314, 1.3764937, 1.393495, 1.4105026, 1.4284993, 1.4455014, 1.462515, 1.4805129, 1.4975198, 1.5145278, 1.5315331, 1.5485371, 1.5655568, 1.5835495, 1.600557, 1.6175622, 1.6345673, 1.6515744, 1.6685906, 1.6865871, 1.7036074, 1.7216546, 1.7396051, 1.7566252, 1.7746162, 1.7916232, 1.8086856, 1.8256917, 1.8436434, 1.8606487, 1.8776556, 1.8946613, 1.9116673, 1.9286712, 1.9456788, 1.9626854, 1.9796923, 1.9967528, 2.0147044], "positions": [[647, 905], [435, 878], [434, 878], [434, 878], [433, 877], [433, 877], [431, 877], [429, 874], [426, 873], [425, 872], [424, 872], [424, 872], [423, 870], [421, 870], [420, 870], [420, 870], [419, 870], [418, 870], [416, 870], [413, 870], [409, 870], [403, 869], [395, 867], [389, 864], [381, 860], [375, 858], [369, 854], [364, 848], [360, 842], [356, 832], [353, 824], [350, 817], [350, 807], [350, 798], [349, 790], [349, 784], [349, 780], [349, 779], [349, 778], [349, 778], [351, 774], [353, 770], [356, 767], [357, 764], [358, 762], [361, 759], [367, 754], [379, 745], [398, 733], [422, 717], [452, 704], [487, 695], [522, 693], [558, 693], [596, 701], [631, 716], [673, 741], [726, 773], [782, 800], [843, 813], [907, 815], [972, 799], [1029, 773], [1081, 739], [1117, 702], [1136, 664], [1138, 632], [1123, 607], [1093, 593], [1056, 593], [1015, 612], [964, 647], [908, 693], [853, 743], [795, 797], [743, 840], [696, 870], [658, 882], [625, 883], [598, 875], [574, 867], [554, 858], [541, 849], [535, 834], [533, 813], [533, 790], [537, 783], [541, 780], [547, 780]]}
